[v0.85][rust] Refactor adl/src/execute/mod.rs by responsibility (Closes #1014)
